+++ python/lsst/ts/observatory/control/mock/atcs_mock.py
@@ -110,8 +110,11 @@ class ATCSMock(BaseGroupMock):
         self.atptg.cmd_azElTarget.callback = self.slew_callback
         self.atptg.cmd_planetTarget.callback = self.slew_callback
         self.atptg.cmd_stopTracking.callback = self.stop_tracking_callback
+        self.atptg.cmd_poriginOffset.callback = self.generic_callback
 
         self.atdome.cmd_moveAzimuth.callback = self.move_dome
+
+        self.atdometrajectory.cmd_setFollowingMode.callback = self.generic_callback
 
         self.ataos_corrections = {
             "m1",